Renamed reorder to mine_variable_order to avoid confusion.
