+++ crates/symbolic/Cargo.toml
@@ -31,8 +31,9 @@ itertools.workspace = true
 log.workspace = true
 oxidd-core.workspace = true
 oxidd-dump.workspace = true
-oxidd.workspace = true
+oxidd-reorder.workspace = true
 oxidd-rules-bdd.workspace = true
+oxidd.workspace = true
 rand.workspace = true
 rustc-hash.workspace = true
 streaming-iterator.workspace = true
