+++ crates/symbolic/src/lib.rs
@@ -14,7 +14,7 @@ mod random_symbolic_lts;
 mod reachability;
 mod reachability_bdd;
 mod refine;
-mod reorder;
+mod mince_variable_order;
 mod sigref;
 mod symbolic_lts;
 mod symbolic_lts_bdd;
@@ -35,7 +35,7 @@ pub use random_symbolic_lts::*;
 pub use reachability::*;
 pub use reachability_bdd::*;
 pub use refine::*;
-pub use reorder::*;
+pub use mince_variable_order::*;
 pub use sigref::*;
 pub use symbolic_lts::*;
 pub use symbolic_lts_bdd::*;
