+++ crates/symbolic/src/mince_variable_order.rs
